Moved files into a new 'data' folder, converted the "please install fcsy" into an f-string, added required packages in "required.txt" file, renamed module "facsim", changed comments to reStructured text instead of epydoc, aligned paragraphs, and changed outfile path to generate fcs file inside of data folder
