+++ requirements.txt
@@ -1,0 +1,4 @@
+numpy~=1.18.1
+pandas~=0.24.2
+fcsy~=0.3.1
+pytest~=5.4.2
